update and fix 2019-3-8
update and fix 2019-3-8

+++ Java/Loon-Neo/src/loon/action/sprite/effect/FadeEffect.java
@@ -66,7 +66,7 @@ public class FadeEffect extends Entity implements BaseEffect {
 		return time;
 	}
 
-	public void setDelay(int delay) {
+	public void setDelay(float delay) {
 		this.time = delay;
 		if (type == TYPE_FADE_IN) {
 			this.currentFrame = this.time;
